+++ src/algoritmos/ordenacao/HeapSort.h
@@ -1,7 +1,7 @@
 #ifndef HEAPSORT_H
 #define HEAPSORT_H
 
-void heapify(unsigned *id, unsigned tam, unsigned indice, unsigned *trocas, unsigned *comparacoes);
-void heapSort(unsigned *id, unsigned tam, unsigned *trocas, unsigned *comparacoes);
+void heapify(unsigned *id, unsigned tam, unsigned indice, unsigned long long *trocas, unsigned long long *comparacoes);
+void heapSort(unsigned *id, unsigned tam, unsigned long long *trocas, unsigned long long *comparacoes);
 
 #endif // HEAPSORT_H
