+++ src/main/cenario1.cpp
@@ -21,7 +21,7 @@ void fluxoCenario1(Review* conjunto, unsigned n, fstream &saida){
     for (unsigned versao = 0; versao < 2; versao++){
 
         // Métricas de desempenho
-        unsigned numComparacoes = 0, numCopias = 0;
+        unsigned long long numComparacoes = 0, numCopias = 0;
 
         switch(versao){
