+++ src/main/cenario2.cpp
@@ -17,7 +17,7 @@ void fluxoCenario2(unsigned* ids, unsigned n, fstream &saida){
     for (unsigned versao = 0; versao < 5; versao++){
 
         // Métricas de desempenho
-        unsigned numComparacoes = 0, numCopias = 0;
+        unsigned long long numComparacoes = 0, numCopias = 0;
 
         switch(versao){
